use macwire instead of cake pattern
