+++ core/amber/build.sbt
@@ -23,7 +23,10 @@ libraryDependencies ++= Seq(
   "com.typesafe.akka" %% "akka-testkit" % akkaVersion,
   "com.typesafe.akka" %% "akka-persistence" % akkaVersion,
   "io.kamon" % "sigar-loader" % "1.6.6-rev002",
-  "com.chuusai" %% "shapeless" % "2.3.3"
+  "com.softwaremill.macwire" %% "macros" % "2.3.6" % "provided",
+  "com.softwaremill.macwire" %% "macrosakka" % "2.3.6" % "provided",
+  "com.softwaremill.macwire" %% "util" % "2.3.6",
+  "com.softwaremill.macwire" %% "proxy" % "2.3.6"
 )
 
 val excludeHadoopJersey = ExclusionRule(organization = "com.sun.jersey")
