+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/WorkerBase.scala
@@ -7,11 +7,12 @@ import edu.uci.ics.amber.engine.common.amberexception.AmberException
 import edu.uci.ics.amber.engine.common.ambermessage.ControlMessage._
 import edu.uci.ics.amber.engine.common.ambermessage.WorkerMessage._
 import edu.uci.ics.amber.engine.common.tuple.ITuple
-import edu.uci.ics.amber.engine.common.ElidableStatement
+import edu.uci.ics.amber.engine.common.{ElidableStatement, IOperatorExecutor}
 import akka.actor.{Actor, ActorLogging, Stash}
 import akka.event.LoggingAdapter
 import akka.util.Timeout
-import edu.uci.ics.amber.engine.architecture.worker.neo.PauseUtil
+import com.softwaremill.macwire.wire
+import edu.uci.ics.amber.engine.architecture.worker.neo.{BatchInput, DataProcessor, PauseUtil, TupleInput, TupleOutput}
 
 import scala.annotation.elidable
 import scala.annotation.elidable.INFO
@@ -20,10 +21,18 @@ import scala.concurrent.ExecutionContext
 import scala.util.control.Breaks
 import scala.concurrent.duration._
 
-abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTransferSupport with PauseUtil {
+abstract class WorkerBase extends Actor with ActorLogging with Stash{
   implicit val ec: ExecutionContext = context.dispatcher
   implicit val timeout: Timeout = 5.seconds
   implicit val logAdapter: LoggingAdapter = log
+
+  var operator: IOperatorExecutor
+
+  lazy val batchInput: BatchInput = wire[BatchInput]
+  lazy val tupleInput: TupleInput = wire[TupleInput]
+  lazy val pauseUtil: PauseUtil = wire[PauseUtil]
+  lazy val tupleOutput: TupleOutput = wire[TupleOutput]
+  lazy val dataProcessor: DataProcessor = wire[DataProcessor]
 
   val receivedFaultedTupleIds: mutable.HashSet[Long] = new mutable.HashSet[Long]()
   val receivedRecoveryInformation: mutable.HashSet[(Long, Long)] =
@@ -39,9 +48,9 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
 
   def onSkipTuple(faultedTuple: FaultedTuple): Unit = {
     if (faultedTuple.isInput) {
-      skippedInputTuples.add(faultedTuple.tuple)
+      tupleOutput.skippedInputTuples.add(faultedTuple.tuple)
     } else {
-      skippedOutputTuples.add(faultedTuple.tuple)
+      tupleOutput.skippedOutputTuples.add(faultedTuple.tuple)
     }
   }
 
@@ -56,7 +65,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
   }
 
   def onPausing(): Unit = {
-    pauseDataTransfer()
+    tupleOutput.pauseDataTransfer()
   }
 
   def onPaused(): Unit = {
@@ -64,7 +73,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
   }
 
   def onResuming(): Unit = {
-    resumeDataTransfer()
+    tupleOutput.resumeDataTransfer()
   }
 
   def onResumed(): Unit = {
@@ -72,17 +81,17 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
   }
 
   def onCompleted(): Unit = {
-    endDataTransfer()
+    tupleOutput.endDataTransfer()
     isCompleted = true
-    resume(PauseUtil.Forced)
+    pauseUtil.resume(PauseUtil.Forced)
     context.parent ! ReportState(WorkerState.Completed)
   }
 
 
   def onBreakpointTriggered(): Unit = {
-    breakpoints.foreach { brk =>
+    dataProcessor.breakpoints.foreach { brk =>
       if (brk.isTriggered)
-        unhandledFaultedTuples(brk.triggeredTupleId) =
+        dataProcessor.unhandledFaultedTuples(brk.triggeredTupleId) =
           new FaultedTuple(brk.triggeredTuple, brk.triggeredTupleId, brk.isInput)
     }
     context.parent ! ReportState(WorkerState.LocalBreakpointTriggered)
@@ -107,21 +116,21 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
   final def allowStashOrReleaseOutput: Receive = {
     case StashOutput =>
       sender ! Ack
-      pauseDataTransfer()
+      tupleOutput.pauseDataTransfer()
     case ReleaseOutput =>
       sender ! Ack
-      resumeDataTransfer()
+      tupleOutput.resumeDataTransfer()
   }
 
   final def allowModifyBreakpoints: Receive = {
     case AssignBreakpoint(bp) =>
       log.info("Assign breakpoint: " + bp.id)
-      registerBreakpoint(bp)
+      dataProcessor.registerBreakpoint(bp)
       sender ! Ack
     case RemoveBreakpoint(id) =>
       log.info("Remove breakpoint: " + id)
       sender ! Ack
-      removeBreakpoint(id)
+      dataProcessor.removeBreakpoint(id)
 
   }
 
@@ -141,7 +150,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
 
   final def allowQueryBreakpoint: Receive = {
     case QueryBreakpoint(id) =>
-      val toReport = breakpoints.find(_.id == id)
+      val toReport = dataProcessor.breakpoints.find(_.id == id)
       if (toReport.isDefined) {
         toReport.get.isReported = true
         context.parent ! ReportedQueriedBreakpoint(toReport.get)
@@ -157,7 +166,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
 
   final def allowQueryTriggeredBreakpoints: Receive = {
     case QueryTriggeredBreakpoints =>
-      val toReport = breakpoints.filter(_.isTriggered)
+      val toReport = dataProcessor.breakpoints.filter(_.isTriggered)
       if (toReport.nonEmpty) {
         toReport.foreach(_.isReported = true)
         sender ! ReportedTriggeredBreakpoints(toReport)
@@ -176,7 +185,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
   final def allowUpdateOutputLinking: Receive = {
     case UpdateOutputLinking(policy, tag, receivers) =>
       sender ! Ack
-      updateOutput(policy, tag, receivers)
+      tupleOutput.updateOutput(policy, tag, receivers)
   }
 
   final def disallowUpdateOutputLinking: Receive = {
@@ -253,7 +262,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
       allowModifyBreakpoints orElse
       disallowQueryTriggeredBreakpoints orElse [Any, Unit] {
       case QueryBreakpoint(id) =>
-        val toReport = breakpoints.find(_.id == id)
+        val toReport = dataProcessor.breakpoints.find(_.id == id)
         if (toReport.isDefined) {
           toReport.get.isReported = true
           context.parent ! ReportedQueriedBreakpoint(toReport.get)
@@ -280,8 +289,8 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
       allowModifyBreakpoints orElse
       disallowQueryTriggeredBreakpoints orElse [Any, Unit] {
       case Resume =>
-        unhandledFaultedTuples.values.foreach(onResumeTuple)
-        unhandledFaultedTuples.clear()
+        dataProcessor.unhandledFaultedTuples.values.foreach(onResumeTuple)
+        dataProcessor.unhandledFaultedTuples.clear()
         onResuming()
         onResumed()
         context.become(running)
@@ -290,21 +299,21 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
         sender ! Ack
         if (!receivedFaultedTupleIds.contains(f.id)) {
           receivedFaultedTupleIds.add(f.id)
-          unhandledFaultedTuples.remove(f.id)
+          dataProcessor.unhandledFaultedTuples.remove(f.id)
           onSkipTuple(f)
         }
       case ModifyTuple(f) =>
         sender ! Ack
         if (!receivedFaultedTupleIds.contains(f.id)) {
           receivedFaultedTupleIds.add(f.id)
-          unhandledFaultedTuples.remove(f.id)
+          dataProcessor.unhandledFaultedTuples.remove(f.id)
           onModifyTuple(f)
         }
       case ResumeTuple(f) =>
         sender ! Ack
         if (!receivedFaultedTupleIds.contains(f.id)) {
           receivedFaultedTupleIds.add(f.id)
-          unhandledFaultedTuples.remove(f.id)
+          dataProcessor.unhandledFaultedTuples.remove(f.id)
           onResumeTuple(f)
         }
       case Pause      => context.parent ! ReportState(WorkerState.Paused)
@@ -314,7 +323,7 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
           WorkerStatistics(WorkerState.Paused, getInputRowCount(), getOutputRowCount())
         )
       case QueryBreakpoint(id) =>
-        val toReport = breakpoints.find(_.id == id)
+        val toReport = dataProcessor.breakpoints.find(_.id == id)
         if (toReport.isDefined) {
           toReport.get.isReported = true
           context.parent ! ReportedQueriedBreakpoint(toReport.get)
@@ -372,8 +381,8 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
       case AssignBreakpoint(bp) =>
         log.info("assign breakpoint: " + bp)
         sender ! Ack
-        registerBreakpoint(bp)
-        if (!breakpoints.exists(_.isDirty)) {
+        dataProcessor.registerBreakpoint(bp)
+        if (!dataProcessor.breakpoints.exists(_.isDirty)) {
           onPaused() //back to paused
           context.unbecome()
           unstashAll()
@@ -381,8 +390,8 @@ abstract class WorkerBase extends Actor with ActorLogging with Stash with DataTr
       case RemoveBreakpoint(id) =>
         log.info("remove breakpoint: " + id)
         sender ! Ack
-        removeBreakpoint(id)
-        if (!breakpoints.exists(_.isDirty)) {
+        dataProcessor.removeBreakpoint(id)
+        if (!dataProcessor.breakpoints.exists(_.isDirty)) {
           onPaused() //back to paused
           context.unbecome()
           unstashAll()
