+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/neo/TupleOutput.scala
@@ -1,6 +1,7 @@
 package edu.uci.ics.amber.engine.architecture.worker.neo
 
+import akka.actor.ActorRef
 import edu.uci.ics.amber.engine.architecture.worker.DataTransferSupport
 
 // use the old data transfer support for now since I don't want to change too much
-trait TupleOutputUtil extends DataTransferSupport
+class TupleOutput(sender:ActorRef) extends DataTransferSupport(sender)
